% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglr-package.R
\docType{package}
\name{anglr-package}
\alias{anglr-package}
\title{Tidy tables for topological spatial data structures.}
\description{
The 'anglr' package show-cases getting past limitations in spatial data by
extending models in the `silicate`` package:
\itemize{
\item coordinates beyond X and Y, or longitude and latitude
\item storing attributes on vertices, primitives, branches (parts), or objects
\item topology and geometry are properly separated
\item spatial data can be properly represented as a graph of spatial primitives
\item polygons as true surfaces, not just glorified lines with a path-filling rule
\item TBD higher dimensional primitives are possible
\item TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction
}
}
\section{I. Creation}{

\tabular{ll}{
\code{\link[silicate]{SC}}, \code{\link[silicate]{PATH}}, \code{\link[silicate]{TRI}}, and
\code{\link[silicate]{ARC}} silicate models are all supported, including the structural forms SC0, TRI0, PATH0 \cr
\code{Spatial} \tab most spatial types can be used directly \cr
\code{\link{DEL}} \tab create a mostly-Delaunay shape-preserving constrained triangulation  \cr
}
}

\section{II. Plotting}{

\tabular{ll}{
\code{\link{globe}} \tab convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form \cr
\code{\link{plot3d.SC}} \tab plot 1D topology in 3D geometry space \cr
\code{\link{plot3d.TRI}} \tab plot 2D topology in 3D geometry space (DEL or TRI) \cr
}
}

\section{III. Options and technicalities}{


There is an option set for the maximum number of triangles that can be
generated by the \code{\link[=DEL]{DEL()}} or \code{\link[=DEL0]{DEL0()}} models when using the \code{max_area}
argument. Inspect the limit with \code{getOption("anglr.max.triangles")} or set a
new limit with \verb{options(anglr.max.triangles = <new limit>)}.

In terms of the \code{\link[RTriangle:triangulate]{RTriangle::triangulate()}} function the \code{max_area} argument
controls and masks the \code{a} argument for RTriangle. It's possible to pass in
values for \code{q}, \code{Y}, \code{j}, \code{D}, \code{S}, \code{V}, and \code{Q} - but we don't recommend
experimenting with these unless you know what they are for. There are no
checks for the limit on the number of triangles that can be created and
several of these do affect that.
}

