% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglr-lines.R, R/anglr-points.R,
%   R/anglr-raster.R, R/anglr-trip.R, R/anglr-vector-classes.R, R/rgl.r
\name{anglr}
\alias{anglr}
\alias{anglr.SpatialMultiPoints}
\alias{anglr.SpatialPoints}
\alias{anglr}
\alias{anglr.RasterLayer}
\alias{anglr}
\alias{anglr.trip}
\alias{anglr.SpatialLines}
\alias{anglr.SpatialPolygons}
\alias{anglr.mesh3d}
\title{Generate primitive-based spatial structures}
\usage{
anglr(x, ...)

\method{anglr}{SpatialMultiPoints}(x, ...)

\method{anglr}{SpatialPoints}(x, ...)

\method{anglr}{RasterLayer}(x, z = x, na.rm = FALSE, ...)

\method{anglr}{trip}(x, ...)

\method{anglr}{SpatialLines}(x, ...)

\method{anglr}{SpatialPolygons}(x, max_area = NULL, ...)

\method{anglr}{mesh3d}(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}

\item{z}{\code{\link[raster]{raster}}, by default \code{x} is used}

\item{na.rm}{remove missing values}

\item{max_area}{maximum area in coordinate system of x, passed to \code{\link[RTriangle]{triangulate}} 'a' argument}

\item{...}{ignored}
}
\value{
a list of tibble data frames, using the gris-map_table model

quad_mesh
}
\description{
Create primitive-based "mesh" structures from various inputs.

Colours not supported, this just gives the viridis palette sequentially.
}
\details{
#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...
}
\section{Warning}{

anglr only checks for presence of triangle centres within
known holes, so this doesn't pick up examples of overlapping areas e.g. 
https://github.com/hypertidy/anglr/issues/39
}

\examples{
## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- anglr(wrld_simpl)
plot(b)
#if (require(rworldxtra)) {

#data(countriesHigh)
#sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
#  "Indonesia", "Malaysia", "Fiji", "Australia")
#a <- subset(countriesHigh, SOVEREIGNT \%in\% sv)
#b7 <- anglr(a, max_area = 0.5)
#plot(globe(b7))
#}
## -----------------------------------------------
## LINES
#l1 <- anglr(as(a, "SpatialLinesDataFrame") )
#plot(l1)
#plot(globe(l1))

#data("flight_tracks", package = "silicate")
#r <- anglr(flight_tracks)
#plot(r)
#rgl::aspect3d(1, 1, 0.001)
#rgl::rglwidget()
library(raster)
w <- raster(volcano)
plot(anglr(w/300))

## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- anglr(dodecahedron3d(col = "cyan"))
octo <- anglr(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)
plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)
bg3d("grey")
}
