% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{as.mesh3d}
\alias{as.mesh3d}
\alias{as.mesh3d.TRI}
\alias{as.mesh3d.default}
\alias{as.mesh3d.TRI0}
\alias{as.mesh3d.matrix}
\title{Mesh3d objects}
\usage{
\method{as.mesh3d}{TRI}(x, z, ..., meshColor = "faces", keep_all = TRUE)

\method{as.mesh3d}{default}(x, ...)

\method{as.mesh3d}{TRI0}(x, ..., meshColor = "faces")

\method{as.mesh3d}{matrix}(x, triangles = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{TRI} or \code{TRI0}}

\item{z}{numeric vector or raster object (see details)}

\item{...}{arguments passed to \code{\link[rgl:tmesh3d]{rgl::tmesh3d()}}}

\item{meshColor}{rule for material properties used for colours (see \link[rgl:tmesh3d]{rgl::tmesh3d})}

\item{keep_all}{whether to keep non-visible triangles}
}
\description{
Methods for the mesh3d type from package rgl
}
\details{
The 'z' argument can be a constant value or a vector of values to be
used for each vertex. Alternatively, it may be a spatial raster object
from which 'z' values are derived. If not set, the vertex 'z_' value
from TRI/TRI0 is used, otherwise z = 0' is assumed.
}
\examples{
sf <- silicate::minimal_mesh
#sf <- silicate::inlandwaters
x <- silicate::TRI(sf)
library(rgl)
clear3d(); plot3d(x); view3d(phi = -10)

# manual face colours (it's not guaranteed that triangle order is native
# within original objects)

clear3d(); plot3d(as.mesh3d(x, material = list(color = rainbow(14))))

mts <- list(color = c("black", "grey")[c(rep(1, 12), c(2, 2))])
clear3d(); plot3d(as.mesh3d(x, material = mts))

## smear by vertices meshColor
mts1 <- list(color = c("black", "grey"))
clear3d(); plot3d(as.mesh3d(x, material = mts1), meshColor = "vertices")

x0 <- silicate::TRI0(sf)
clear3d(); plot3d(x0); view3d(phi = -10)

# (TRI0 - it *is* guaranteed that triangle order is native)
clear3d(); plot3d(as.mesh3d(x0,  material = list(color = rainbow(14))))

## arbitrarily drape polygons over raster
r <- raster::setExtent(raster::raster(volcano), raster::extent(-0.1, 1.1, -0.1, 1.1))
clear3d();shade3d(as.mesh3d(DEL(silicate::minimal_mesh, max_area = 0.001), z =r))
aspect3d(1, 1, 0.5)
}
