% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_3d}
\alias{auto_3d}
\title{Auto aspect ratio}
\usage{
auto_3d(x = 1, y = 1, z = 1, keep_xy = TRUE, exag = TRUE,
  silent = FALSE)
}
\arguments{
\item{x}{exaggeration for x}

\item{y}{exaggeration for y}

\item{z}{exaggeration for z}

\item{keep_xy}{should xy be forced to maintain their current aspect ratio}

\item{exag}{should the x, y, z factors be applied, set to \code{FALSE} to ignore}
}
\value{
the output of \code{rgl::par3d} invisibly
}
\description{
Automatically modify the aspect ratio of a scene to
rescale drastically different data ranges into something more
pleasing.
}
\details{
This is typically used to rescale data in different units, for example longitude and
latitude in degrees and elevation in metres.
}
\examples{
topo <- copy_down(silicate::SC(simpleworld), gebco1)
plot3d(topo);  rgl::rglwidget()

auto_3d(z = 4); rgl::rglwidget()
}
