% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_down.R
\name{copy_down}
\alias{copy_down}
\alias{copy_down}
\alias{copy_down.sc}
\alias{copy_down}
\alias{copy_down.PATH}
\alias{copy_down}
\alias{copy_down.TRI}
\title{Copy down values to vertices}
\usage{
copy_down(x, z = NULL, ..., .id = "z_")

\method{copy_down}{sc}(x, z = NULL, ..., .id = "z_")

\method{copy_down}{PATH}(x, z = NULL, ..., .id = "z_")

\method{copy_down}{TRI}(x, z = NULL, ..., .id = "z_")
}
\arguments{
\item{x}{a silicate model}

\item{z}{object specifying values to copy down, a vector of values, a column name, a raster (see details)}

\item{...}{currently ignored}

\item{.id}{character value, the name of the resulting column in the vertices, default is "z_"}
}
\value{
silicate model with vertex values copied to vertices
}
\description{
Copy down provides ways to transfer object level data values to
vertex level.
}
\details{
Various methods are used depending on the second argument `z`.

If z is a raster (`BasicRaster`) a numeric value for each vertex is found by bilinear
interpolation using `raster::extract(raster, vertex, method = "bilinear")`. Vertices
are transformed into the space used by the raster if possible. (WIP ... Otherwise a warning is issued if 
there's not overlap ... WIP)

If z is a character value, that column is taken from the object table.

The `.id` argument must be character and exist as a column name in the object table.

If z is a vector it's simply copied down.

No checking is done on the type of the result, and so there's nothing to stop the use of the recyling rule
to expand out values, and nothing to stop the use of non numeric values being copied down. It's your model.
}
\examples{

}
