% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot3d.R
\name{dot3d}
\alias{dot3d}
\alias{dot3d.sf}
\alias{dot3d.matrix}
\alias{dot3d.BasicRaster}
\alias{dot3d.sc}
\alias{dot3d.SC}
\alias{dot3d.SC0}
\title{Draw points in 3d}
\usage{
\method{dot3d}{sf}(x, ...)

\method{dot3d}{matrix}(x, ...)

\method{dot3d}{BasicRaster}(x, ...)

\method{dot3d}{sc}(x, ...)

\method{dot3d}{SC}(x, ...)

\method{dot3d}{SC0}(x, ...)
}
\arguments{
\item{x}{sc, sp, sf, raster, trip, or any other model understood by anglr/silicate}

\item{...}{
  Arguments passed on to \code{\link[rgl:as.mesh3d]{rgl::as.mesh3d}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Draw points with rgl from any mesh-alike object.
}
\examples{
dot3d(cad_tas)
dot3d(volcano, size = 10)
auto_3d(z = 14)


rgl::clear3d()
## from ?persp
y <- x <- seq(-10, 10, length= 80)
z <- outer(x, y,
     function(x, y) {
        r <- sqrt(x^2+y^2); 10 * sin(r)/r
        })
dot3d(z)
auto_3d()
dot3d(raster::raster(volcano), size = 10)
auto_3d(z = 14)
\donttest{
dot3d(silicate::SC(cad_tas))
auto_3d(z = 14)
}
}
