% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.r
\name{globe}
\alias{globe}
\alias{globe.default}
\title{Convert map coordinates to Geocentric (XYZ) coordinates.}
\usage{
globe(x, ...)

\method{globe}{default}(x, gproj = "+proj=geocent +ellps=WGS84", ...)
}
\arguments{
\item{x}{list of tibbles, in \code{\link{mesh}} form}

\item{...}{arguments to methods (none used)}

\item{gproj}{Geocentric PROJ.4 string, defaults to WGS84}
}
\value{
mesh object with vertices table modified
}
\description{
Convert map coordinates to Geocentric (XYZ) coordinates.
}
\examples{
library(maptools)
data(wrld_simpl)
g <- globe(mesh(as(wrld_simpl, "SpatialLinesDataFrame")))
plot(g, lwd = 3)
}

