% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{mesh3d}
\alias{mesh3d}
\alias{as.mesh3d.QUAD}
\alias{as.mesh3d.default}
\alias{as.mesh3d.TRI}
\title{Mesh3d objects}
\usage{
\method{as.mesh3d}{QUAD}(x, ...)

\method{as.mesh3d}{default}(x, ...)

\method{as.mesh3d}{TRI}(x, ...)
}
\description{
Methods for the mesh3d type from package rgl
}
