% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp3d.R
\name{persp3d}
\alias{persp3d}
\alias{persp3d.TRI}
\alias{persp3d.TRI0}
\alias{persp3d.BasicRaster}
\alias{persp3d.DEL}
\alias{persp3d.DEL0}
\alias{persp3d.QUAD}
\alias{persp3d.matrix}
\alias{persp3d.sf}
\alias{persp3d.sfc}
\alias{persp3d.Spatial}
\alias{persp3d.triangulation}
\title{persp3d}
\usage{
\method{persp3d}{TRI}(x, ..., add = FALSE)

\method{persp3d}{TRI0}(x, ..., add = FALSE)

\method{persp3d}{BasicRaster}(x, ..., add = FALSE)

\method{persp3d}{DEL}(x, ..., add = FALSE)

\method{persp3d}{DEL0}(x, ..., add = FALSE)

\method{persp3d}{QUAD}(x, ..., add = FALSE)

\method{persp3d}{matrix}(x, ..., add = FALSE)

\method{persp3d}{sf}(x, ..., add = FALSE)

\method{persp3d}{sfc}(x, ..., add = FALSE)

\method{persp3d}{Spatial}(x, ..., add = FALSE)

\method{persp3d}{triangulation}(x, ..., add = FALSE)
}
\arguments{
\item{x}{sp, sf, raster, trip, or any other model understood by anglr/silicate}

\item{...}{pass \link[rgl:material]{material3d properties} to rgl}

\item{add}{add to existing plot or start a new one (the default)}
}
\description{
Plot surface
}
\seealso{
\link{plot3d} \link{as.mesh3d} \link{wire3d} \link{dot3d} \link{shade3d}
}
