% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot-rangl}
\alias{plot-rangl}
\alias{plot.trimesh}
\alias{plot}
\alias{plot-rangl}
\alias{plot.linemesh}
\alias{plot-rangl}
\alias{plot.pointmesh}
\title{Plot objects in OpenGL}
\usage{
\method{plot}{trimesh}(x, ..., add = FALSE)

\method{plot}{linemesh}(x, ..., add = FALSE)

\method{plot}{pointmesh}(x, ..., add = FALSE)
}
\arguments{
\item{x}{object from \code{\link{rangl}}}

\item{...}{args for underlying plotting}

\item{add}{add to existing plot if exists}
}
\value{
the rgl mesh3d object, invisibly
}
\description{
Plot using the \code{\link[rgl]{rgl-package}}.
}
\details{
The data structures from \code{\link{rangl}} are converted to their analogous forms
used by the \code{\link[rgl]{rgl}} package and plotted. These plot methods return
the rgl form invisibly.
}
