% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.QUAD}
\alias{plot.QUAD}
\title{Plot QUAD}
\usage{
\method{plot}{QUAD}(x, ..., add = FALSE)
}
\arguments{
\item{x}{}

\item{...}{}

\item{add}{}
}
\description{
Plot a QUAD mesh.
}
\examples{
plot(QUAD(raster::raster(volcano)))
}
