% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangl-raster.r
\name{plot.quad_mesh}
\alias{plot.quad_mesh}
\title{Title}
\usage{
\method{plot}{quad_mesh}(x, ...)
}
\arguments{
\item{x}{quad_mesh}

\item{...}{args passed to rgl plot}
}
\value{
qmesh
}
\description{
Title
}
\examples{
example(rangl.RasterLayer)
plot(w)
}
