% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.QUAD}
\alias{plot.linemesh}
\alias{plot.pointmesh}
\title{Plot QUAD}
\usage{
\method{plot}{QUAD}(x, ..., add = FALSE)

\method{plot}{linemesh}(x, ..., add = FALSE)

\method{plot}{pointmesh}(x, ..., add = FALSE)
}
\arguments{
\item{x}{a QUAD object}

\item{...}{passed to set up plot when silicate mesh is plotted}

\item{add}{add to the current plot?}
}
\value{
returns the data plotted in the form used, invisibly
}
\description{
Plot a QUAD mesh.
}
\examples{
plot(QUAD(raster::raster(volcano)))
}
