% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_mesh.r
\name{plot.trimesh}
\alias{plot.trimesh}
\title{plot the triangles in the tables}
\usage{
\method{plot}{trimesh}(x, ...)
}
\arguments{
\item{x}{object from tri_mesh}

\item{...}{args for underlying plotting}
}
\value{
the rgl mesh3d object, invisibly
}
\description{
plot
}
\examples{
example(tri_mesh)
if(exists("b")) { 
 ##plot(b)
 }
}

