% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{plot3d.SC}
\alias{plot3d.SC}
\title{3D object plot}
\usage{
\method{plot3d}{SC}(x, ..., add = FALSE)
}
\arguments{
\item{x}{silicate model, SC, TRI, ARC, or PATH}

\item{...}{}

\item{add}{}
}
\value{
vertices and segment indices, invisibly
}
\description{
3D object plot
}
\examples{
library(silicate)
x <- SC(sf::read_sf(system.file("shape", "nc.shp", package = "sf")) \%>\% 
   mutate(color_ = rainbow(100)))
plot3d(x); rglwidget()
worldz <- QUAD(gebco1)
## an easy way to exaggerate z is to reduce the radius of the globe
plot3d(globe(worldz, gproj = "+proj=geocent +a=10000"))
}
