% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUAD-model.R
\name{QUAD}
\alias{QUAD}
\alias{QUAD.matrix}
\alias{QUAD.BasicRaster}
\title{QUAD model}
\usage{
QUAD(x, ...)

\method{QUAD}{matrix}(x, ...)

\method{QUAD}{BasicRaster}(x, ...)
}
\arguments{
\item{x}{raster alike, or a matrix}

\item{...}{ignored}
}
\value{
QUAD model
}
\description{
The QUAD model
}
\details{
This is lazy, object table only stores the raster extent.

The 'color_' idiom works, but must be put on the '$quad' table.
Very WIP.
}
