% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangl-polygons.r
\name{tri_mesh}
\alias{mesh}
\alias{tri_mesh}
\title{Deprecated from rangl}
\usage{
tri_mesh(x, ...)

mesh(x, ...)
}
\arguments{
\item{x}{nothing}

\item{...}{ignored}
}
\description{
Deprecated from rangl
}

