% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangl-raster.r
\name{rangl.RasterLayer}
\alias{rangl.RasterLayer}
\title{Raster rangl}
\usage{
\method{rangl}{RasterLayer}(x, z = x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{\code{\link[raster]{raster}}}

\item{z}{\code{\link[raster]{raster}}, by default \code{x} is used}

\item{na.rm}{remove missing values}

\item{...}{unused}
}
\value{
quad_mesh
}
\description{
Colours not supported, this just gives the virids palette sequentially.
}
\examples{
library(raster)
w <- raster(volcano)
plot(rangl(w/300))

}

