% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_extent.R
\name{sf_extent}
\alias{sf_extent}
\title{Extent of simple features}
\usage{
sf_extent(x)
}
\arguments{
\item{x}{sf object}
}
\value{
a raster Extent
}
\description{
Get Extent of sf objects
}
\details{
This function exists to avoid raster needing the sf package. We don't need
GDAL, GEOS, and all the other fun stuff to find four numbers that describe
coordinate range, so this is a workaround.
}
\examples{
sf_extent(silicate::inlandwaters)
}
