% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglr-package.R
\name{silicate-models}
\alias{silicate-models}
\alias{PATH}
\alias{PATH0}
\alias{SC}
\alias{SC0}
\alias{TRI0}
\alias{ARC}
\title{silicate models}
\arguments{
\item{x}{any data format understood by the model}

\item{...}{unused}
}
\description{
The anglr functions \code{\link[=DEL]{DEL()}}, \code{\link[=DEL0]{DEL0()}} and \code{\link[=QUAD]{QUAD()}} extend the models
of the silicate package. In particular \code{DEL()} and \code{DEL0()} are high
quality Delaunay-constrained triangulation meshers analogous to the
ear-cutting algorithms used by \code{\link[silicate:TRI]{silicate::TRI()}} and \code{\link[silicate:TRI0]{silicate::TRI0()}}.
}
\details{
All models in silicate are imported by anglr and re-exported.
}
\seealso{
\code{\link[=DEL]{DEL()}} \code{\link[=DEL0]{DEL0()}} \code{\link[=QUAD]{QUAD()}} \code{\link[silicate:PATH]{silicate::PATH()}} \code{\link[silicate:TRI]{silicate::TRI()}} \code{\link[silicate:ARC]{silicate::ARC()}} \code{\link[silicate:SC]{silicate::SC()}} \code{\link[silicate:PATH0]{silicate::PATH0()}} \code{\link[silicate:TRI0]{silicate::TRI0()}} \code{\link[silicate:SC0]{silicate::SC0()}}
}
