% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglr-package.R
\docType{data}
\name{simpleworld}
\alias{simpleworld}
\title{simple world}
\description{
A simple polygon map of world sovereign countries, a modified copy of
the rnaturalearth counties110 (see data-raw/simpleworld.R for details).
}
\examples{
DEL(simpleworld[1:10, ])
}
