% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadToTriangle.R
\name{TRI.QUAD}
\alias{TRI.QUAD}
\title{TRI model extensions}
\usage{
TRI.QUAD(x, ...)
}
\arguments{
\item{x}{QUAD}
}
\description{
TRI model from silicate is extended by methods for QUAD.
}
\examples{
library(anglr)
library(raster)
r <- raster(volcano)
a <- QUAD(r)
x <- TRI(a)

mesh <- DEL(x, max_area = 0.1)
}
