% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_mesh.r
\name{tri_mesh}
\alias{tri_mesh}
\alias{tri_mesh.SpatialPolygons}
\title{Generate triangle mesh}
\usage{
tri_mesh(x, ...)

\method{tri_mesh}{SpatialPolygons}(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}
}
\value{
a list of tibble data frames, using the gris-map_table model
}
\description{
Create triangle mesh structures from various inputs.
}
\details{
Methods exist for SpatialPolygons, ...
}
\examples{
if (require(rworldxtra)) {

data(countriesHigh)
sv <- "Australia"
a <- subset(countriesHigh, SOVEREIGNT == sv)
b <- tri_mesh(a)
}
}

