% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimesh-package.r
\docType{package}
\name{trimesh}
\alias{trimesh}
\alias{trimesh-package}
\title{Triangle mesh}
\description{
Triangle mesh
}
\section{I. Creation}{

\tabular{ll}{
 \code{\link{tri_mesh}} \tab create a triangulated mesh as tidy set of tables  \cr
 }
}

\section{II. Plotting}{

\tabular{ll}{
 \code{\link{plot.trimesh}} \tab plot in 3D, by default just in the plane \cr
}
}

