% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire3d.R
\name{wire3d}
\alias{wire3d}
\alias{wire3d.sc}
\alias{wire3d.TRI}
\alias{wire3d.TRI0}
\alias{wire3d.DEL}
\alias{wire3d.DEL0}
\alias{wire3d.QUAD}
\alias{wire3d.matrix}
\alias{wire3d.sf}
\alias{wire3d.Spatial}
\alias{wire3d.triangulation}
\alias{wire3d.trip}
\alias{wire3d.BasicRaster}
\title{wire3}
\usage{
\method{wire3d}{sc}(x, ...)

\method{wire3d}{TRI}(x, ...)

\method{wire3d}{TRI0}(x, ...)

\method{wire3d}{DEL}(x, ...)

\method{wire3d}{DEL0}(x, ...)

\method{wire3d}{QUAD}(x, ...)

\method{wire3d}{matrix}(x, ...)

\method{wire3d}{sf}(x, ...)

\method{wire3d}{Spatial}(x, ...)

\method{wire3d}{triangulation}(x, ...)

\method{wire3d}{trip}(x, ...)

\method{wire3d}{BasicRaster}(x, ...)
}
\arguments{
\item{x}{sp, sf, raster, trip, or any other model understood by anglr/silicate}

\item{...}{pass \link[rgl:rgl.material]{material3d properties} to rgl}
}
\description{
plot lines from mesh
}
\seealso{
plot3d as.mesh3d persp3d dot3d
}
