% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglr-package.R
\docType{package}
\name{anglr-package}
\alias{anglr-package}
\title{Tidy tables for topological spatial data structures.}
\description{
The 'anglr' package helps transcend the following general limitations:
\itemize{
\item coordinates beyond X and Y, or longitude and latitude
\item storing attributes on vertices, primitives, branches (parts), or objects
\item topology and geometry are properly separated
\item spatial data can be properly represented as a graph of spatial primitives
\item polygons as true surfaces, not just glorified lines with a path-filling rule
\item TBD higher dimensional primitives are possible
\item TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction
}
}
\section{I. Creation}{

\tabular{ll}{
\code{\link{anglr}} \tab create a anglr table set from various input types  \cr
}
}

\section{II. Plotting}{

\tabular{ll}{
\code{\link{globe}} \tab convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form \cr
\code{\link{plot.trimesh}} \tab plot 2D topology in 3D geometry space \cr
\code{\link{plot.linemesh}} \tab plot 1D topology in 3D geometry space \cr
\code{\link{plot.pointmesh}} \tab plot 0D topology in 3D geometry space \cr
}
}

