% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{as.mesh3d}
\alias{as.mesh3d}
\alias{as.mesh3d.QUAD}
\alias{as.mesh3d.default}
\alias{as.mesh3d.TRI}
\title{Mesh3d objects}
\usage{
\method{as.mesh3d}{QUAD}(x, ...)

\method{as.mesh3d}{default}(x, ...)

\method{as.mesh3d}{TRI}(x, ...)
}
\arguments{
\item{x}{
For the generic, \code{x} is the object to convert.
For the default method, \code{x}, \code{y} and \code{z} are
coordinates. Any reasonable way of defining the coordinates is
acceptable. See the function \code{\link{xyz.coords}} for details.
}

\item{...}{
Material properties to pass to \code{\link{tmesh3d}} or \code{\link{qmesh3d}}.
}
}
\description{
Methods for the mesh3d type from package rgl
}
