% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEL0-model.R
\name{DEL0}
\alias{DEL0}
\alias{DEL0.DEL}
\alias{DEL0.default}
\alias{DEL0.SC}
\alias{DEL0.SC0}
\alias{DEL0.TRI}
\alias{DEL0.TRI0}
\alias{DEL0.ARC}
\alias{DEL0.PATH}
\alias{DEL0.PATH0}
\title{Convert object to a constrained-Delaunay triangulation}
\usage{
DEL0(x, ..., max_area = NULL)

\method{DEL0}{DEL}(x, ..., max_area = NULL)

\method{DEL0}{default}(x, ..., max_area = NULL)

\method{DEL0}{SC}(x, ..., max_area = NULL)

\method{DEL0}{SC0}(x, ..., max_area = NULL)

\method{DEL0}{TRI}(x, ..., max_area = NULL)

\method{DEL0}{TRI0}(x, ..., max_area = NULL)

\method{DEL0}{ARC}(x, ..., max_area = NULL)

\method{DEL0}{PATH}(x, ..., max_area = NULL)

\method{DEL0}{PATH0}(x, ..., max_area = NULL)
}
\arguments{
\item{x}{object of class \link{PATH0} or understood by \code{\link[=PATH0]{PATH0()}}}

\item{...}{ignored}

\item{max_area}{the maximum area of a triangle}
}
\value{
\link[=DEL0]{DEL0 class}
}
\description{
This \emph{structural-form} Delaunay-based triangulation model is analogous to the
\link[silicate:TRI]{TRI()} model in the silicate package and formally extends the
class of that model. A primitives-based shape-constrained triangulation. The
Delaunay model is the \emph{mostly Delaunay} scheme used by the provable-quality
meshers.
}
\details{
This is a more compact form of the \emph{relational-form} \code{\link[=DEL]{DEL()}} model.
}
\seealso{
\link{DEL}
}
