% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEL0-model.R
\name{DEL0}
\alias{DEL0}
\alias{DEL0.default}
\alias{DEL0.PATH0}
\title{DEL0}
\usage{
DEL0(x, ..., max_area = NULL)

\method{DEL0}{default}(x, ..., max_area = NULL)

\method{DEL0}{PATH0}(x, ..., max_area = NULL)
}
\arguments{
\item{x}{object of class \link{PATH0} or understood by \code{\link[=PATH0]{PATH0()}}}

\item{...}{ignored}

\item{max_area}{the maximum area of a triangle}
}
\value{
DEL0 class
}
\description{
Structural form of Delaunay constrained triangulation
}
\details{
More compact form of \code{\link[=DEL]{DEL()}} model.
}
\examples{
a <- DEL0(cad_tas)
plot(a)
}
