% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_mesh.r
\name{globe}
\alias{globe}
\alias{globe.trimesh}
\title{Title}
\usage{
globe(x, ...)

\method{globe}{trimesh}(x, halo = FALSE, ..., rad = 1)
}
\arguments{
\item{x}{}

\item{...}{}

\item{halo}{}

\item{rad}{}
}
\description{
Title
}

