% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_mesh.r
\name{globe}
\alias{globe}
\alias{globe.trimesh}
\title{Title}
\usage{
globe(x, ...)

\method{globe}{trimesh}(x, halo = FALSE, ..., rad = 1)
}
\arguments{
\item{x}{object from tri_mesh}

\item{...}{passed to plot}

\item{halo}{show the radius}

\item{rad}{radius}
}
\value{
the mesh object, invisibly
}
\description{
Title
}
\examples{
example(tri_mesh)
if(exists("b")) { 
 ##globe(b, halo = TRUE)
 }
}

