% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.r
\name{globe}
\alias{globe}
\alias{globe.default}
\title{Convert map coordinates to Geocentric (XYZ) coordinates.}
\usage{
globe(x, ...)

\method{globe}{default}(x, gproj = "+proj=geocent +ellps=WGS84", ...)
}
\arguments{
\item{x}{list of tibbles, in \code{\link{anglr}} form}

\item{...}{arguments to methods (none used)}

\item{gproj}{Geocentric PROJ.4 string, defaults to WGS84}
}
\value{
anglr object with vertices table modified
}
\description{
Convert map coordinates to Geocentric (XYZ) coordinates.
}
\examples{
data(simpleworld)
g <- globe(silicate::PATH(as(simpleworld, "SpatialLinesDataFrame")))
if (interactive()) {
 plot(g, lwd = 3)
 }
}
