% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.r
\name{mesh}
\alias{mesh}
\alias{mesh.SpatialLines}
\title{Generate primitives mesh}
\usage{
mesh(x, ...)

\method{mesh}{SpatialLines}(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}
}
\value{
a list of tibble data frames, using the gris-map_table model
}
\description{
Create primitives mesh structures from various inputs.
}
\details{
Methods exist for ...
}

