% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.r, R/points.r, R/rgl.r, R/tri_mesh.r
\name{mesh}
\alias{mesh}
\alias{mesh.SpatialLines}
\alias{mesh.SpatialMultiPoints}
\alias{mesh.SpatialPoints}
\alias{mesh.SpatialPolygons}
\alias{mesh.mesh3d}
\title{Generate primitives mesh}
\usage{
mesh(x, ...)

\method{mesh}{SpatialLines}(x, ...)

\method{mesh}{SpatialMultiPoints}(x, ...)

\method{mesh}{SpatialPoints}(x, ...)

\method{mesh}{mesh3d}(x, ...)

\method{mesh}{SpatialPolygons}(x, max_area = NULL, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}

\item{max_area}{maximum area in coordinate system of x, passed to \code{\link[RTriangle]{triangulate}} 'a' argument}
}
\value{
a list of tibble data frames, using the gris-map_table model
}
\description{
Create primitives mesh structures from various inputs.
}
\details{
#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...
}
\examples{
## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- mesh(wrld_simpl)
plot(b)
if (require(rworldxtra)) {

data(countriesHigh)
sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
 "Indonesia", "Malaysia", "Fiji", "Australia")
a <- subset(countriesHigh, SOVEREIGNT \%in\% sv)
b7 <- mesh(a, max_area = 0.5)
plot(globe(b7))
}
## -----------------------------------------------
## LINES
l1 <- mesh(as(a, "SpatialLinesDataFrame") )
plot(l1)
plot(globe(l1))
## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- mesh(dodecahedron3d(col = "cyan"))
octo <- mesh(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)
plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)
bg3d("grey")
}

