% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot-rangl}
\alias{plot}
\alias{plot-rangl}
\alias{plot.linemesh}
\alias{plot.pointmesh}
\alias{plot.quad_mesh}
\alias{plot.trimesh}
\title{Plot objects in OpenGL}
\usage{
\method{plot}{trimesh}(x, ...)

\method{plot}{linemesh}(x, ...)

\method{plot}{pointmesh}(x, ...)

\method{plot}{quad_mesh}(x, ...)
}
\arguments{
\item{x}{object from \code{\link{rangl}}}

\item{...}{args for underlying plotting}
}
\value{
the rgl mesh3d object, invisibly
}
\description{
Plot using the \code{\link[rgl]{rgl-package}}.
}
\details{
The data structures from \code{\link{rangl}} are converted to their analogous forms
used by the \code{\link[rgl]{rgl}} package and plotted. These plot methods return
the rgl form invisibly.
}
\examples{
#example(rangl)
#if(exists("b")) { 
#  plot(b)
#  }
example(rangl.RasterLayer)
plot(w)
}

