% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot-rangl}
\alias{plot}
\alias{plot-rangl}
\alias{plot.linemesh}
\alias{plot.pointmesh}
\alias{plot.trimesh}
\title{plot the triangles in the tables}
\usage{
plot.trimesh(x, ...)

plot.linemesh(x, ...)

plot.pointmesh(x, ...)
}
\arguments{
\item{x}{object from \code{\link{rangl}}}

\item{...}{args for underlying plotting}
}
\value{
the rgl mesh3d object, invisibly
}
\description{
plot
}
\examples{
example(rangl)
if(exists("b")) { 
 plot(b)
 }
}

