% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.quad_mesh}
\alias{plot.quad_mesh}
\title{Title}
\usage{
\method{plot}{quad_mesh}(x, ..., add = FALSE)
}
\arguments{
\item{x}{quad_mesh}

\item{...}{args passed to rgl plot}

\item{add}{reset plot?}
}
\value{
qmesh
}
\description{
Title
}
\seealso{
anglr.RasterLayer
}
