% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUAD-model.R
\name{QUAD}
\alias{QUAD}
\alias{QUAD}
\alias{QUAD.BasicRaster}
\title{QUAD model}
\usage{
QUAD(x, ...)

\method{QUAD}{BasicRaster}(x, ...)
}
\arguments{
\item{x}{raster alike}

\item{...}{ignored}
}
\value{
QUAD model
}
\description{
The QUAD model
}
\examples{
QUAD(raster::raster(volcano))
}
