% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangl-package.r
\docType{package}
\name{rangl}
\alias{rangl}
\alias{rangl-package}
\title{Tidy tables for topological, triangulated, and other spatial data structures.}
\description{
Tidy tables for topological, triangulated, and other spatial data structures.
}
\section{I. Creation}{

\tabular{ll}{
 \code{\link{mesh}} \tab create a triangulated mesh as tidy set of tables  \cr
 }
}

\section{II. Plotting}{

\tabular{ll}{
 \code{\link{globe}} \tab convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form \cr
 \code{\link{plot.trimesh}} \tab plot 2D topology in 3D \cr
 \code{\link{plot.trimesh}} \tab plot 1D topology in 3D \cr
}
}

