% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangl-lines.r, R/rangl-package.r, R/rangl-points.r, R/rangl-polygons.r, R/rangl-trip.r, R/rgl.r
\docType{package}
\name{rangl}
\alias{rangl}
\alias{rangl-package}
\alias{rangl.SpatialLines}
\alias{rangl.SpatialMultiPoints}
\alias{rangl.SpatialPoints}
\alias{rangl.SpatialPolygons}
\alias{rangl.mesh3d}
\alias{rangl.trip}
\title{Generate primitive-based spatial structures}
\usage{
rangl(x, ...)

\method{rangl}{SpatialLines}(x, ...)

\method{rangl}{SpatialMultiPoints}(x, ...)

\method{rangl}{SpatialPoints}(x, ...)

\method{rangl}{SpatialPolygons}(x, max_area = NULL, ...)

\method{rangl}{trip}(x, ...)

\method{rangl}{mesh3d}(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}

\item{max_area}{maximum area in coordinate system of x, passed to \code{\link[RTriangle]{triangulate}} 'a' argument}
}
\value{
a list of tibble data frames, using the gris-map_table model
}
\description{
Create primitive-based "mesh" structures from various inputs.

The 'rangl' package helps transcend the following problems in R and GIS generally:
}
\details{
#' Methods exist for SpatialPolygons, SpatialLines, rgl mesh3d(triangle) ...

\itemize{
 \item coordinates beyond X and Y, or longitude and latitude
 \item attributes stored where appropriate, on vertices, primitives, branches, or objects
 \item topology and geometry are not conflated
 \item spatial data can be properly represented as a graph of spatial primitives
 \item polygons as true surfaces, not just glorified lines with a path-filling rule
 \item TBD higher dimensional primitives are possible
 \item TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction 
}
}
\section{I. Creation}{

\tabular{ll}{
 \code{\link{rangl}} \tab create a rangl table set from various input types  \cr
 }
}

\section{II. Plotting}{

\tabular{ll}{
 \code{\link{globe}} \tab convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form \cr
 \code{\link{plot.trimesh}} \tab plot 2D topology in 3D geometry space \cr
 \code{\link{plot.linemesh}} \tab plot 1D topology in 3D geometry space \cr
 \code{\link{plot.pointmesh}} \tab plot 0D topology in 3D geometry space \cr
}
}
\examples{
## -----------------------------------------------
## POLYGONS
library(maptools)
data(wrld_simpl)
b <- rangl(wrld_simpl)
plot(b)
if (require(rworldxtra)) {

data(countriesHigh)
sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
 "Indonesia", "Malaysia", "Fiji", "Australia")
a <- subset(countriesHigh, SOVEREIGNT \%in\% sv)
b7 <- rangl(a, max_area = 0.5)
plot(globe(b7))
}
## -----------------------------------------------
## LINES
l1 <- rangl(as(a, "SpatialLinesDataFrame") )
plot(l1)
plot(globe(l1))
## -----------------------------------------------
## RGL TRIANGLES
library(rgl)

dod <- rangl(dodecahedron3d(col = "cyan"))
octo <- rangl(translate3d(octahedron3d(col = "blue"), 6, 0, 0))
plot(dod, col = viridis::viridis(5)[1], alpha = 0.3)
plot(octo, col = viridis::viridis(5)[5], alpha = 0.3)
bg3d("grey")
}

