% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade3d.R
\name{shade3d}
\alias{shade3d}
\alias{shade3d.TRI}
\alias{shade3d.TRI0}
\alias{shade3d.DEL}
\alias{shade3d.DEL0}
\alias{shade3d.QUAD}
\alias{shade3d.matrix}
\alias{shade3d.sf}
\alias{shade3d.Spatial}
\title{shade3d}
\usage{
\method{shade3d}{TRI}(x, ..., add = FALSE)

\method{shade3d}{TRI0}(x, ..., add = FALSE)

\method{shade3d}{DEL}(x, ..., add = FALSE)

\method{shade3d}{DEL0}(x, ..., add = FALSE)

\method{shade3d}{QUAD}(x, ..., add = FALSE)

\method{shade3d}{matrix}(x, ..., add = FALSE)

\method{shade3d}{sf}(x, ..., add = FALSE)

\method{shade3d}{Spatial}(x, ..., add = FALSE)
}
\arguments{
\item{x}{sp, sf, raster, trip, or any other model understood by anglr/silicate}

\item{...}{pass \link[rgl:rgl.material]{material3d properties} to rgl}

\item{add}{add to existing plot or start a new one (the default)}
}
\description{
Plot surface
}
\seealso{
\link{plot3d} \link{as.mesh3d} \link{persp3d} \link{dot3d} \link{wire3d}
}
