% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadToTriangle.R
\name{TRI.QUAD}
\alias{TRI.QUAD}
\title{TRI model extensions}
\usage{
\method{TRI}{QUAD}(x, ...)
}
\arguments{
\item{x}{QUAD}
}
\description{
TRI model from silicate is extended by methods for QUAD.
}
\examples{
library(anglr)
library(raster)
r <- setExtent(raster(volcano), extent(0, nrow(volcano), 0, ncol(volcano)))
a <- QUAD(r)
x <- copy_down(TRI(a), r)
nrow(x$vertex)
mesh <- DEL(x, max_area = .2)
mesh <- copy_down(mesh, r)
nrow(mesh$vertex)
}
