% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadToTriangle.R
\name{TRI.QUAD}
\alias{TRI.QUAD}
\alias{TRI}
\title{TRI model extensions}
\usage{
\method{TRI}{QUAD}(x, ...)
}
\arguments{
\item{x}{QUAD}

\item{...}{reserved}
}
\value{
a TRI model, as per silicate package
}
\description{
TRI model from silicate is extended by methods for QUAD.
}
\examples{
library(anglr)
library(raster)
v <- volcano[1:10, 1:6]
r <- setExtent(raster(v), extent(0, nrow(v), 0, ncol(v)))
a <- QUAD(r)
x <- copy_down(TRI(a), r)
nrow(x$vertex)
sc <- silicate::SC(x)
#mesh <- DEL(sc, max_area = .2)
#mesh <- copy_down(mesh, r)
#nrow(mesh$vertex)
}
