% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.r, R/tri_mesh.r
\name{tri_mesh.mesh3d}
\alias{tri_mesh}
\alias{tri_mesh.SpatialPolygons}
\alias{tri_mesh.mesh3d}
\title{Generate triangle mesh}
\usage{
\method{tri_mesh}{mesh3d}(x)

tri_mesh(x, ...)

\method{tri_mesh}{SpatialPolygons}(x, max_area = NULL, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}

\item{max_area}{maximum area in coordinate system of x, passed to \code{\link[RTriangle]{triangulate}} 'a' argument}
}
\value{
a list of tibble data frames, using the gris-map_table model
}
\description{
Create triangle mesh structures from various inputs.
}
\details{
Methods exist for SpatialPolygons, rgl mesh3d(triangle) ...
}
\examples{
if (require(rworldxtra)) {

data(countriesHigh)
sv <- c("New Zealand", "Antarctica", "Papua New Guinea",
 "Indonesia", "Malaysia", "Fiji", "Australia")
a <- subset(countriesHigh, SOVEREIGNT \%in\% sv)
b <- tri_mesh(a)
#b <- tri_mesh(a, max_area = 0.1)
}

library(maptools)
data(wrld_simpl)
b <- tri_mesh(wrld_simpl)
}

