% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire3d.R
\name{wire3d}
\alias{wire3d}
\alias{wire3d.sc}
\alias{wire3d.TRI}
\alias{wire3d.TRI0}
\alias{wire3d.DEL}
\alias{wire3d.DEL0}
\alias{wire3d.QUAD}
\alias{wire3d.matrix}
\alias{wire3d.sf}
\alias{wire3d.Spatial}
\alias{wire3d.triangulation}
\alias{wire3d.trip}
\alias{wire3d.BasicRaster}
\title{Draw a mesh as line segments in 3D}
\usage{
wire3d.sc(x, ...)

wire3d.TRI(x, ...)

wire3d.TRI0(x, ...)

wire3d.DEL(x, ...)

wire3d.DEL0(x, ...)

wire3d.QUAD(x, ...)

wire3d.matrix(x, ...)

wire3d.sf(x, ...)

wire3d.Spatial(x, ...)

wire3d.triangulation(x, ...)

wire3d.trip(x, ...)

wire3d.BasicRaster(x, ...)
}
\arguments{
\item{x}{sp, sf, raster, trip, or any other model understood by anglr/silicate}

\item{...}{pass \link[rgl:material]{material3d properties} to rgl}
}
\description{
Draw line segments with rgl from any \link[rgl:mesh3d]{shape3d} classed object. Produces
a 3D scatterplot like that produced by \code{\link[rgl:plot3d]{rgl::plot3d()}}, but from a mesh-alike
object.
}
\details{
Objects that are not explicitly surfaces will be triangulated in order to produce the
mesh. Whether this is a good idea or not is an open question.

It is not currently \emph{technically defined or clear} how colour properties are
mapped to line segments by default ... there is a problem of what property to
use from features that share the same vertex or edge, and we have put that
aside and erred on the side of inaccuracy in favour of getting a pretty plot
(hopefully). (Properties that come later - lower rows - win, I think.
}
\seealso{
\link{plot3d} \link{as.mesh3d} \link{persp3d} \link{dot3d} \link{shade3d}
}
